Dev 0.0.11 创建数据库初始化方法（DataBaseInit::main()）

+++ src/main/java/io/mlqs/es/legal/db/MarriageRegistrationRecordEntity.java
@@ -39,5 +39,5 @@ public class MarriageRegistrationRecordEntity extends LegalRecordEntity {
      * 向量化
      */
     @Field(type = FieldType.Dense_Vector, dims = 1024)
-    private float[] vector;
+    private double[] vector;
 }
